API Moneda

+++ BackBilletera/MiBilleteraWebApi/Controllers/UsuariosController.cs
@@ -1,14 +1,15 @@
 ﻿using MiBilleteraWebApi.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace MiBilleteraWebApi.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("api/usuarios")]
     [ApiController]
-    public class UsuarioController : ControllerBase
+    public class UsuariosController : ControllerBase
     {
 
         // GET: api/<UsuarioController>
@@ -40,7 +41,7 @@ namespace MiBilleteraWebApi.Controllers
                 var existeUsuario = db.Usuarios.FirstOrDefault(x => x.Dni == usuario.Dni);
                 if(existeUsuario != null)
                 {
-                    return NotFound();
+                    return BadRequest("El usuario que quiere registrar, ya se encuentra registrado anteriormente.");
                 }
                 db.Add(usuario);
                 db.SaveChanges();
@@ -48,26 +49,8 @@ namespace MiBilleteraWebApi.Controllers
             }
         }
 
-        // PUT api/<UsuarioController>/5
-        [HttpPut("{id}")]
-        public ActionResult Put(int id, Usuario usuario)
-        {
-            using (var db = new Pil2022Context())
-            {
-                var existeUsuario = db.Usuarios.FirstOrDefault(x => x.IdUsuario == id);
-                if(existeUsuario!= null)
-                {
-                    db.Update(usuario);
-                    db.SaveChanges();
-                    return Ok();
-                }
-                return NotFound();
-            }
-        }
-
-        // DELETE api/<UsuarioController>/5
-        [HttpDelete("{id}")]
-        public ActionResult Delete(int id)
+        [HttpPost("baja/{id:int}")]
+        public ActionResult PostEstado(int id)
         {
             using (var db = new Pil2022Context())
             {
@@ -76,9 +59,51 @@ namespace MiBilleteraWebApi.Controllers
                 {
                     return NotFound();
                 }
-                db.Remove(new Usuario() { IdUsuario = id});
+                existeUsuario.FehcaBaja = DateTime.Now;
                 db.SaveChanges();
                 return Ok();
+            }
+        }
+
+
+        // PUT api/<UsuarioController>/5
+        [HttpPut("{id:int}")]
+        public ActionResult Put(int id, Usuario usuario)
+        {
+            using (var db = new Pil2022Context())
+            {
+                if(usuario.IdUsuario != id)
+                {
+                    return NotFound();
+                }
+                db.Update(usuario);
+                db.SaveChanges();
+                return Ok();
+            }
+        }
+
+        // DELETE api/<UsuarioController>/5
+        [HttpDelete("{id:int}")]
+        public ActionResult Delete(int id)
+        {
+            using (var db = new Pil2022Context())
+            {
+                try
+                {
+                    var existeUsuario = db.Usuarios.Any(x => x.IdUsuario == id);
+                    if (existeUsuario == null)
+                    {
+                        return NotFound();
+                    }
+                    db.Remove(new Usuario() { IdUsuario = id });
+                    db.SaveChangesAsync();
+                    return Ok();
+                }
+                catch (Exception ex)
+                {
+                    throw;
+                }
+
             }
         }
     }
